+++ src/main/scala/ai/starlake/extract/ExtractJDBCSchema.scala
@@ -45,7 +45,9 @@ class ExtractJDBCSchema(schemaHandler: SchemaHandler) extends Extract with LazyL
         YamlSerializer.deserializeJDBCSchemas(content, config.extractConfig)
       val connectionOptions = jdbcSchemas.connectionRef
         .map(settings.appConfig.connections(_).options)
-        .getOrElse(jdbcSchemas.connection)
+        .getOrElse(
+          throw new Exception(s"No connectionRef found. Please check your connectionRef property")
+        )
       implicit val forkJoinTaskSupport: Option[ForkJoinTaskSupport] =
         ExtractUtils.createForkSupport(config.parallelism)
       ExtractUtils.makeParallel(jdbcSchemas.jdbcSchemas).foreach { jdbcSchema =>