+++ src/main/scala/ai/starlake/schema/model/WriteMode.scala
@@ -20,7 +20,7 @@
 
 package ai.starlake.schema.model
 
-import ai.starlake.schema.model.WriteMode.{APPEND, ERROR_IF_EXISTS, IGNORE, OVERWRITE}
+import ai.starlake.schema.model.WriteMode.{APPEND, ERROR_IF_EXISTS, IGNORE, MERGE, OVERWRITE}
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.annotation.{JsonDeserialize, JsonSerialize}
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer
@@ -42,6 +42,7 @@ sealed case class WriteMode(value: String) {
     this match {
       case OVERWRITE       => SaveMode.Overwrite
       case APPEND          => SaveMode.Append
+      case MERGE           => SaveMode.Append
       case ERROR_IF_EXISTS => SaveMode.ErrorIfExists
       case IGNORE          => SaveMode.Ignore
       case _ =>
@@ -54,10 +55,10 @@ object WriteMode {
 
   def fromString(value: String): WriteMode = {
     value.toUpperCase() match {
-      case "OVERWRITE"       => WriteMode.OVERWRITE
-      case "APPEND"          => WriteMode.APPEND
-      case "ERROR_IF_EXISTS" => WriteMode.ERROR_IF_EXISTS
-      case "IGNORE"          => WriteMode.IGNORE
+      case "OVERWRITE"        => WriteMode.OVERWRITE
+      case "APPEND" | "MERGE" => WriteMode.APPEND
+      case "ERROR_IF_EXISTS"  => WriteMode.ERROR_IF_EXISTS
+      case "IGNORE"           => WriteMode.IGNORE
       case _ =>
         throw new Exception(s"Invalid Write Mode try one of ${writes}")
     }
@@ -66,6 +67,8 @@ object WriteMode {
   object OVERWRITE extends WriteMode("OVERWRITE")
 
   object APPEND extends WriteMode("APPEND")
+
+  object MERGE extends WriteMode("MERGE")
 
   object ERROR_IF_EXISTS extends WriteMode("ERROR_IF_EXISTS")